linter: move dirwalk-related code to workspace pkg
The FS traversal does not depend on the linter package
details, so we can move it somewhere else.

workspace package is intended to be used as a toolbox
for working with the PHP project (its files, settings, etc).
