+++ src/langsrv/refs.go
@@ -14,6 +14,7 @@ import (
 	"github.com/VKCOM/noverify/src/solver"
 	"github.com/VKCOM/noverify/src/state"
 	"github.com/VKCOM/noverify/src/vscode"
+	"github.com/VKCOM/noverify/src/workspace"
 	"go.lsp.dev/uri"
 )
 
@@ -192,8 +193,8 @@ func refPosition(filename string, pos *position.Position) vscode.Location {
 type parseFn func(filename string, rootNode ir.Node, contents []byte, parser *php7.Parser) []vscode.Location
 
 func findReferences(substr string, parse parseFn) []vscode.Location {
-	cb := linter.ReadFilenames(linter.AnalysisFiles, nil)
-	ch := make(chan linter.FileInfo)
+	cb := workspace.ReadFilenames(linter.AnalysisFiles, nil)
+	ch := make(chan workspace.FileInfo)
 	go func() {
 		cb(ch)
 		close(ch)
