+++ src/linter/parser.go
@@ -4,10 +4,7 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
-	"os"
-	"path/filepath"
 	"regexp"
 	dbg "runtime/debug"
 	"strings"
@@ -15,8 +12,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/karrick/godirwalk"
-	"github.com/monochromegane/go-gitignore"
 	"github.com/quasilyte/regex/syntax"
 
 	"github.com/VKCOM/noverify/src/git"
@@ -28,50 +23,8 @@ import (
 	"github.com/VKCOM/noverify/src/php/parser/php7"
 	"github.com/VKCOM/noverify/src/quickfix"
 	"github.com/VKCOM/noverify/src/rules"
+	"github.com/VKCOM/noverify/src/workspace"
 )
-
-type FileInfo struct {
-	Filename   string
-	Contents   []byte
-	LineRanges []git.LineRange
-}
-
-func isPHPExtension(filename string) bool {
-	fileExt := filepath.Ext(filename)
-	if fileExt == "" {
-		return false
-	}
-
-	fileExt = fileExt[1:] // cut "." in the beginning
-
-	for _, ext := range PHPExtensions {
-		if fileExt == ext {
-			return true
-		}
-	}
-
-	return false
-}
-
-func makePHPExtensionSuffixes() [][]byte {
-	res := make([][]byte, 0, len(PHPExtensions))
-	for _, ext := range PHPExtensions {
-		res = append(res, []byte("."+ext))
-	}
-	return res
-}
-
-func isPHPExtensionBytes(filename []byte, suffixes [][]byte) bool {
-	for _, suffix := range suffixes {
-		if bytes.HasSuffix(filename, suffix) {
-			return true
-		}
-	}
-
-	return false
-}
-
-type ReadCallback func(ch chan FileInfo)
 
 // ParseContents parses specified contents (or file) and returns *RootWalker.
 // Function does not update global meta.
@@ -229,306 +182,8 @@ func DebugMessage(msg string, args ...interface{}) {
 	}
 }
 
-// ParseGitignoreFromDir tries to parse a gitignore file at path/.gitignore.
-// If no such file exists, <nil, nil> is returned.
-func ParseGitignoreFromDir(path string) (gitignore.IgnoreMatcher, error) {
-	f, err := os.Open(filepath.Join(path, ".gitignore"))
-	switch {
-	case os.IsNotExist(err):
-		return nil, nil // No gitignore file, not an error
-	case err != nil:
-		return nil, err // Some unexpected error (e.g. access failure)
-	}
-	defer f.Close()
-	matcher := gitignore.NewGitIgnoreFromReader(path, f)
-	return matcher, nil
-}
-
-func readFilenames(ch chan<- FileInfo, filename string, filter *FilenameFilter) {
-	absFilename, err := filepath.Abs(filename)
-	if err == nil {
-		filename = absFilename
-	}
-
-	if filter == nil {
-		// No-op filter that doesn't track gitignore files.
-		filter = &FilenameFilter{}
-	}
-
-	// If we use stat here, it will return file info of an entry
-	// pointed by a symlink (if filename is a link).
-	// lstat is required for a symlink test below to succeed.
-	// If we ever want to permit top-level (CLI args) symlinks,
-	// caller should resolve them to a files that are pointed by them.
-	st, err := os.Lstat(filename)
-	if err != nil {
-		log.Fatalf("Could not stat file %s: %s", filename, err.Error())
-	}
-	if st.Mode()&os.ModeSymlink != 0 {
-		// filepath.Walk does not follow symlinks, but it does
-		// accept it as a root argument without an error.
-		// godirwalk.Walk can traverse symlinks with FollowSymbolicLinks=true,
-		// but we don't use it. It will give an error if root is
-		// a symlink, so we avoid calling Walk() on them.
-		return
-	}
-
-	if !st.IsDir() {
-		if filter.IgnoreFile(filename) {
-			return
-		}
-
-		ch <- FileInfo{Filename: filename}
-		return
-	}
-
-	// Start with a sentinel "" path to make last(gitignorePaths) safe
-	// without a length check.
-	gitignorePaths := []string{""}
-
-	walkOptions := &godirwalk.Options{
-		Unsorted: true,
-
-		Callback: func(path string, de *godirwalk.Dirent) error {
-			if de.IsDir() {
-				if filter.IgnoreDir(path) {
-					return filepath.SkipDir
-				}
-				// During indexing phase and with -gitignore=false
-				// we don't want to do extra FS operations.
-				if !filter.GitignoreIsEnabled() {
-					return nil
-				}
-
-				matcher, err := ParseGitignoreFromDir(path)
-				if err != nil {
-					linterError(path, "read .gitignore: %v", err)
-				}
-				if matcher != nil {
-					gitignorePaths = append(gitignorePaths, path)
-					filter.GitignorePush(path, matcher)
-				}
-				return nil
-			}
-
-			if !isPHPExtension(path) {
-				return nil
-			}
-			if filter.IgnoreFile(path) {
-				return nil
-			}
-
-			ch <- FileInfo{Filename: path}
-			return nil
-		},
-	}
-
-	if filter.GitignoreIsEnabled() {
-		walkOptions.PostChildrenCallback = func(path string, de *godirwalk.Dirent) error {
-			topGitignorePath := gitignorePaths[len(gitignorePaths)-1]
-			if topGitignorePath == path {
-				gitignorePaths = gitignorePaths[:len(gitignorePaths)-1]
-				filter.GitignorePop(path)
-			}
-			return nil
-		}
-	}
-
-	if err := godirwalk.Walk(filename, walkOptions); err != nil {
-		log.Fatalf("Could not walk filepath %s (%v)", filename, err)
-	}
-}
-
-// ReadFilenames returns callback that reads filenames into channel
-func ReadFilenames(filenames []string, filter *FilenameFilter) ReadCallback {
-	return func(ch chan FileInfo) {
-		for _, filename := range filenames {
-			readFilenames(ch, filename, filter)
-		}
-	}
-}
-
-// ReadChangesFromWorkTree returns callback that reads files from workTree dir that are changed
-func ReadChangesFromWorkTree(dir string, changes []git.Change) ReadCallback {
-	return func(ch chan FileInfo) {
-		for _, c := range changes {
-			if c.Type == git.Deleted {
-				continue
-			}
-
-			if !isPHPExtension(c.NewName) {
-				continue
-			}
-
-			filename := filepath.Join(dir, c.NewName)
-
-			contents, err := ioutil.ReadFile(filename)
-			if err != nil {
-				log.Fatalf("Could not read file %s: %s", filename, err.Error())
-			}
-
-			ch <- FileInfo{
-				Filename: filename,
-				Contents: contents,
-			}
-		}
-	}
-}
-
-// ReadFilesFromGit parses file contents in the specified commit
-func ReadFilesFromGit(repo, commitSHA1 string, ignoreRegex *regexp.Regexp) ReadCallback {
-	catter, err := git.NewCatter(repo)
-	if err != nil {
-		log.Fatalf("Could not start catter: %s", err.Error())
-	}
-
-	tree, err := git.GetTreeSHA1(catter, commitSHA1)
-	if err != nil {
-		log.Fatalf("Could not get tree sha1: %s", err.Error())
-	}
-
-	suffixes := makePHPExtensionSuffixes()
-
-	return func(ch chan FileInfo) {
-		start := time.Now()
-		idx := 0
-
-		err = catter.Walk(
-			"",
-			tree,
-			func(filename []byte) bool {
-				return isPHPExtensionBytes(filename, suffixes)
-			},
-			func(filename string, contents []byte) {
-				idx++
-				if time.Since(start) >= 2*time.Second {
-					start = time.Now()
-					action := "Indexed"
-					if meta.IsIndexingComplete() {
-						action = "Analyzed"
-					}
-					log.Printf("%s %d files from git", action, idx)
-				}
-
-				if ignoreRegex != nil && ignoreRegex.MatchString(filename) {
-					return
-				}
-
-				ch <- FileInfo{
-					Filename: filename,
-					Contents: contents,
-				}
-			},
-		)
-
-		if err != nil {
-			log.Fatalf("Could not walk: %s", err.Error())
-		}
-	}
-}
-
-// ReadOldFilesFromGit parses file contents in the specified commit, the old version
-func ReadOldFilesFromGit(repo, commitSHA1 string, changes []git.Change) ReadCallback {
-	changedMap := make(map[string][]git.LineRange, len(changes))
-	for _, ch := range changes {
-		if ch.Type == git.Added {
-			continue
-		}
-		changedMap[ch.OldName] = append(changedMap[ch.OldName], ch.OldLineRanges...)
-	}
-
-	catter, err := git.NewCatter(repo)
-	if err != nil {
-		log.Fatalf("Could not start catter: %s", err.Error())
-	}
-
-	tree, err := git.GetTreeSHA1(catter, commitSHA1)
-	if err != nil {
-		log.Fatalf("Could not get tree sha1: %s", err.Error())
-	}
-
-	suffixes := makePHPExtensionSuffixes()
-
-	return func(ch chan FileInfo) {
-		err = catter.Walk(
-			"",
-			tree,
-			func(filename []byte) bool {
-				if !isPHPExtensionBytes(filename, suffixes) {
-					return false
-				}
-
-				_, ok := changedMap[string(filename)]
-				return ok
-			},
-			func(filename string, contents []byte) {
-				ch <- FileInfo{
-					Filename:   filename,
-					Contents:   contents,
-					LineRanges: changedMap[filename],
-				}
-			},
-		)
-
-		if err != nil {
-			log.Fatalf("Could not walk: %s", err.Error())
-		}
-	}
-}
-
-// ReadFilesFromGitWithChanges parses file contents in the specified commit, but only specified ranges
-func ReadFilesFromGitWithChanges(repo, commitSHA1 string, changes []git.Change) ReadCallback {
-	changedMap := make(map[string][]git.LineRange, len(changes))
-	for _, ch := range changes {
-		if ch.Type == git.Deleted {
-			// TODO: actually support deletes too
-			continue
-		}
-
-		changedMap[ch.NewName] = append(changedMap[ch.NewName], ch.LineRanges...)
-	}
-
-	catter, err := git.NewCatter(repo)
-	if err != nil {
-		log.Fatalf("Could not start catter: %s", err.Error())
-	}
-
-	tree, err := git.GetTreeSHA1(catter, commitSHA1)
-	if err != nil {
-		log.Fatalf("Could not get tree sha1: %s", err.Error())
-	}
-
-	suffixes := makePHPExtensionSuffixes()
-
-	return func(ch chan FileInfo) {
-		err = catter.Walk(
-			"",
-			tree,
-			func(filename []byte) bool {
-				if !isPHPExtensionBytes(filename, suffixes) {
-					return false
-				}
-
-				_, ok := changedMap[string(filename)]
-				return ok
-			},
-			func(filename string, contents []byte) {
-				ch <- FileInfo{
-					Filename:   filename,
-					Contents:   contents,
-					LineRanges: changedMap[filename],
-				}
-			},
-		)
-
-		if err != nil {
-			log.Fatalf("Could not walk: %s", err.Error())
-		}
-	}
-}
-
 // ParseFilenames is used to do initial parsing of files.
-func ParseFilenames(readFileNamesFunc ReadCallback, allowDisabled *regexp.Regexp) []*Report {
+func ParseFilenames(readFileNamesFunc workspace.ReadCallback, allowDisabled *regexp.Regexp) []*Report {
 	start := time.Now()
 	defer func() {
 		lintdebug.Send("Processing time: %s", time.Since(start))
@@ -543,7 +198,7 @@ func ParseFilenames(readFileNamesFunc ReadCallback, allowDisabled *regexp.Regexp
 
 	lintdebug.Send("Parsing using %d cores", MaxConcurrency)
 
-	filenamesCh := make(chan FileInfo, 512)
+	filenamesCh := make(chan workspace.FileInfo, 512)
 
 	go func() {
 		readFileNamesFunc(filenamesCh)
@@ -574,7 +229,7 @@ func ParseFilenames(readFileNamesFunc ReadCallback, allowDisabled *regexp.Regexp
 	return allReports
 }
 
-func doParseFile(f FileInfo, needReports bool, allowDisabled *regexp.Regexp) (reports []*Report) {
+func doParseFile(f workspace.FileInfo, needReports bool, allowDisabled *regexp.Regexp) (reports []*Report) {
 	var err error
 
 	if DebugParseDuration > 0 {
@@ -604,7 +259,7 @@ func doParseFile(f FileInfo, needReports bool, allowDisabled *regexp.Regexp) (re
 	return reports
 }
 
-func InitStubs(readFileNamesFunc ReadCallback) {
+func InitStubs(readFileNamesFunc workspace.ReadCallback) {
 	meta.SetLoadingStubs(true)
 	ParseFilenames(readFileNamesFunc, nil)
 	meta.Info.InitStubs()
@@ -613,5 +268,5 @@ func InitStubs(readFileNamesFunc ReadCallback) {
 
 // InitStubsFromDir parses directory with PHPStorm stubs which has all internal PHP classes and functions declared.
 func InitStubsFromDir(dir string) {
-	InitStubs(ReadFilenames([]string{dir}, nil))
+	InitStubs(workspace.ReadFilenames([]string{dir}, nil))
 }
