+++ src/tests/exprtype/exprtype_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/VKCOM/noverify/src/linter"
 	"github.com/VKCOM/noverify/src/linttest"
 	"github.com/VKCOM/noverify/src/meta"
+	"github.com/VKCOM/noverify/src/workspace"
 )
 
 // Tests in this file make it less likely that type solving will break
@@ -2161,8 +2162,8 @@ exprtype(Roo::$d, "string");
 func runExprTypeTest(t *testing.T, params *exprTypeTestParams) {
 	meta.ResetInfo()
 	if params.stubs != "" {
-		linter.InitStubs(func(ch chan linter.FileInfo) {
-			ch <- linter.FileInfo{
+		linter.InitStubs(func(ch chan workspace.FileInfo) {
+			ch <- workspace.FileInfo{
 				Filename: "stubs.php",
 				Contents: []byte(params.stubs),
 			}
